requirements.txt
pillow
pystray
PySimpleGUI
psutil
PyWinCtl 